% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_expand.R
\name{dm_expand}
\alias{dm_expand}
\title{Expand a design matrix to include pairwise interactions terms and polynomial terms.}
\usage{
dm_expand(LME = DM, mpo = 4)
}
\arguments{
\item{LME}{The design matrix, each row is an observation, each column is a covariate.  The covariates should include linear main effects only.}

\item{mpo}{the maximum polynomial order to include.}
}
\value{
The expanded design matrix.
}
\description{
\code{dm_expand} expands the supplied design matrix of linear main effects terms for covariates to include columns for all possible pairwise interactions of covariates and columns for polynomial terms for all covariates up to the polynomial order specified to \code{mpo}.  The additional columns for interaction terms are named following the convention where the column named 'first.covariate.name_X_second.covariate.name' contains the products of the observations in 'first.covariate.name' and 'second.covariate.name'.  The additional columns for polynomial terms are named following the convention where the polynomial order is appended to the name of the consituent covariate such that the column named 'covariate.name.2' contains the values for the quadratic (polynomial order 2) terms for the covariate called 'covariate.name'. Note: to avoid one of the consitutent covariates dominating the values of the interaction term all linear main effects terms are recentered to have means of zero and L1 norms of one prior to calculating the pairwise interactions.  Similarly to avoid extremeley large magnitude values in the column for higher order polynomial terms these are also calculate from linear main effects terms that have been recentered to have means of zero and L1 norms of one prior to calculating the pairwise interactions.
}
\examples{
# Expand the design matrix DM: 
X.P4I2 <- dm_expand(LME = DM, mpo = 4)
}

