% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_filter.R
\name{dm_filter}
\alias{dm_filter}
\title{Filter a design matrix to reduce its collinearity.}
\usage{
dm_filter(DM = DM, prefer = c("band1", "band2", "band3", "band4", "band5",
  "band6", "NDVI"), mpccm = 0.8, na.action = c("Ignore", "Halt"))
}
\arguments{
\item{DM}{The design matrix, each row is an observation, each column is a covariate.}

\item{prefer}{An ordered character vector of covariates names.  The first element will be retained by the filtering above all other covariates with which it is correlated. The second element will be retained after the first etc.}

\item{mpccm}{The maximum permitted correlation coefficient magnitude between remaining covariate pairs post filtering.}

\item{na.action}{What to do if there are NAs among the elements of the design matrix supplied to `DM'.}
}
\value{
The output is a two element list consisting of the named elements \code{drop} and \code{FDM}.  The first element of this list, \code{drop}, is a character vector containing the names of all the covariates that were filtered from the design matrix due having correlation coefficient magnitude greater than the cut off value with covariates that have been retained in the filtered design matrix.  The second element of this list, \code{FDM}, is the design matrix that has been filtered to enforce the specified maximum permitted correlation coefficient between remaining covariates.
}
\description{
\code{dm_filter} Filters the provided design matrix to reduce its collinearity.  All pairs of covariates that have correlation coefficient magnitudes greater than the value specified to \code{mpccm} are extracted.  These overtly correlated pairs of covariates are then searched to identify which pairs contain the covariates named in the first element of the character vector \code{prefer}.  All covariates that are overtly correlated with the first element of the list of covariates to prefer are then filtered from the design matrix.  This proceedure is then repeated for each subsequent member of the list of covariates to preferentially retain.  Once this proceedure has been repeated for each covariate on the list of covariates to preferentially retain a simplified version of the filtering is conducted complete the filtering of the design matrix.  A covariate is choosen at random from the first pair remaining on the list of overtly correlated pairs of covariates and any covariates overtly correlated with this covariate are filtered from the design matrix.  This proceedure is then repeated until there are no longer any pairs of covariates among those included in the filtered design matrix which have correlation coefficient magnitudes greater than the value supplied to \code{mpccm}.
}
\examples{
# Filter the design matrix DM:
X.F <- dm_filter(DM = DM, prefer = c('band1', 'band2', 'band3', 'band4', 'band5', 'band6', 'NDVI'), mpccm = 0.8, na.action = 'Ignore')

The covariates that have been filtered from the design matrix:
X.F[['drop']]

The filtered design matrix:
X.F[['FDM']]
}

